+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -623,6 +623,35 @@ class SentryOptionsTest {
     }
 
     @Test
+    fun `when options are initialized, metrics is disabled by default`() {
+        assertFalse(SentryOptions().isEnableMetrics)
+        assertFalse(SentryOptions().isEnableDefaultTagsForMetrics)
+        assertFalse(SentryOptions().isEnableSpanLocalMetricAggregation)
+    }
+
+    @Test
+    fun `when metrics is enabled, getters reflect that`() {
+        val options = SentryOptions().apply {
+            isEnableMetrics = true
+        }
+        assertTrue(options.isEnableMetrics)
+        assertTrue(options.isEnableDefaultTagsForMetrics)
+        assertTrue(options.isEnableSpanLocalMetricAggregation)
+    }
+
+    @Test
+    fun `when metric settings are flipped, getters reflect that`() {
+        val options = SentryOptions().apply {
+            isEnableMetrics = true
+            isEnableDefaultTagsForMetrics = false
+            isEnableSpanLocalMetricAggregation = false
+        }
+        assertTrue(options.isEnableMetrics)
+        assertFalse(options.isEnableDefaultTagsForMetrics)
+        assertFalse(options.isEnableSpanLocalMetricAggregation)
+    }
+
+    @Test
     fun `existing cron defaults are not overridden if not present in external options`() {
         val options = SentryOptions().apply {
             cron = SentryOptions.Cron().apply {