+++ sentry/src/test/java/io/sentry/metrics/MetricsApiTest.kt
@@ -11,16 +11,30 @@ import kotlin.test.Test
 
 class MetricsApiTest {
 
+    class Fixture {
+        val aggregator = mock<IMetricsAggregator>()
+        val localMetricsAggregator = mock<LocalMetricsAggregator>()
+
+        fun getSut(defaultTags: Map<String, String> = emptyMap()): MetricsApi {
+            val localAggregator = localMetricsAggregator
+
+            return MetricsApi(object : IMetricsInterface {
+                override fun getMetricsAggregator(): IMetricsAggregator {
+                    return aggregator
+                }
+
+                override fun getLocalMetricsAggregator(): LocalMetricsAggregator? = localAggregator
+
+                override fun getDefaultTagsForMetrics(): Map<String, String> = defaultTags
+            })
+        }
+    }
+
+    val fixture = Fixture()
+
     @Test
     fun `default timestamp is provided`() {
-        val aggregator = mock<IMetricsAggregator>()
-        val api = MetricsApi(object : IMetricsInterface {
-            override fun getMetricsAggregator(): IMetricsAggregator {
-                return aggregator
-            }
-
-            override fun getDefaultTagsForMetrics(): Map<String, String> = emptyMap()
-        })
+        val api = fixture.getSut()
 
         api.increment("name", 1.0, null, null, null)
         api.set("name", 1, null, null, null)
@@ -28,62 +42,60 @@ class MetricsApiTest {
         api.gauge("name", 1.0, null, null, null)
         api.distribution("name", 1.0, null, null, null)
 
-        verify(aggregator).increment(
+        verify(fixture.aggregator).increment(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             any(),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).set(
+        verify(fixture.aggregator).set(
             anyOrNull(),
             eq(1),
             anyOrNull(),
             anyOrNull(),
             any(),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).set(
+        verify(fixture.aggregator).set(
             anyOrNull(),
             eq("string"),
             anyOrNull(),
             anyOrNull(),
             any(),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).gauge(
+        verify(fixture.aggregator).gauge(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             any(),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).distribution(
+        verify(fixture.aggregator).distribution(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             any(),
+            anyOrNull(),
             anyOrNull()
         )
     }
 
     @Test
     fun `timestamp is not overwritten`() {
-        val aggregator = mock<IMetricsAggregator>()
-        val api = MetricsApi(object : IMetricsInterface {
-            override fun getMetricsAggregator(): IMetricsAggregator {
-                return aggregator
-            }
-
-            override fun getDefaultTagsForMetrics(): Map<String, String> = emptyMap()
-        })
+        val api = fixture.getSut()
 
         api.increment("name", 1.0, null, null, 1234)
         api.set("name", 1, null, null, 1234)
@@ -91,70 +103,69 @@ class MetricsApiTest {
         api.gauge("name", 1.0, null, null, 1234)
         api.distribution("name", 1.0, null, null, 1234)
 
-        verify(aggregator).increment(
+        verify(fixture.aggregator).increment(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             eq(1234),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).set(
+        verify(fixture.aggregator).set(
             anyOrNull(),
             eq(1),
             anyOrNull(),
             anyOrNull(),
             eq(1234),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).set(
+        verify(fixture.aggregator).set(
             anyOrNull(),
             eq("string"),
             anyOrNull(),
             anyOrNull(),
             eq(1234),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).gauge(
+        verify(fixture.aggregator).gauge(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             eq(1234),
+            anyOrNull(),
             anyOrNull()
         )
 
-        verify(aggregator).distribution(
+        verify(fixture.aggregator).distribution(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
             eq(1234),
+            anyOrNull(),
             anyOrNull()
         )
     }
 
     @Test
     fun `tags are enriched with default tags`() {
-        val aggregator = mock<IMetricsAggregator>()
-        val api = MetricsApi(object : IMetricsInterface {
-            override fun getMetricsAggregator(): IMetricsAggregator {
-                return aggregator
-            }
+        val api = fixture.getSut(
+            defaultTags = mapOf(
+                "release" to "1.0",
+                "environment" to "prod"
+            )
+        )
 
-            override fun getDefaultTagsForMetrics(): Map<String, String> {
-                return mapOf(
-                    "release" to "1.0",
-                    "environment" to "prod"
-                )
-            }
-        })
         api.increment("name", 1.0, null, mapOf("a" to "b"))
 
-        verify(aggregator).increment(
+        verify(fixture.aggregator).increment(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
@@ -166,25 +177,20 @@ class MetricsApiTest {
                 )
             ),
             anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
 
     @Test
     fun `existing environment and release tags are not overwritten`() {
-        val aggregator = mock<IMetricsAggregator>()
-        val api = MetricsApi(object : IMetricsInterface {
-            override fun getMetricsAggregator(): IMetricsAggregator {
-                return aggregator
-            }
+        val api = fixture.getSut(
+            defaultTags = mapOf(
+                "release" to "1.0",
+                "environment" to "prod"
+            )
+        )
 
-            override fun getDefaultTagsForMetrics(): Map<String, String> {
-                return mapOf(
-                    "release" to "1.0",
-                    "environment" to "prod"
-                )
-            }
-        })
         api.increment(
             "name",
             1.0,
@@ -195,7 +201,7 @@ class MetricsApiTest {
             )
         )
 
-        verify(aggregator).increment(
+        verify(fixture.aggregator).increment(
             anyOrNull(),
             anyOrNull(),
             anyOrNull(),
@@ -206,7 +212,80 @@ class MetricsApiTest {
                 )
             ),
             anyOrNull(),
+            anyOrNull(),
             anyOrNull()
+        )
+    }
+
+    @Test
+    fun `local aggregator is provided to aggregator`() {
+        val api = fixture.getSut()
+
+        api.increment("increment")
+        verify(fixture.aggregator).increment(
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            eq(fixture.localMetricsAggregator)
+        )
+
+        api.set("set", 1)
+        verify(fixture.aggregator).set(
+            anyOrNull(),
+            eq(1),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            eq(fixture.localMetricsAggregator)
+        )
+
+        api.set("set", "string")
+        verify(fixture.aggregator).set(
+            anyOrNull(),
+            eq("string"),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            eq(fixture.localMetricsAggregator)
+        )
+
+        api.gauge("gauge", 1.0)
+        verify(fixture.aggregator).gauge(
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            eq(fixture.localMetricsAggregator)
+        )
+
+        api.distribution("distribution", 1.0)
+        verify(fixture.aggregator).distribution(
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            eq(fixture.localMetricsAggregator)
+        )
+
+        api.timing("timing") {
+            // no-op
+        }
+        verify(fixture.aggregator).timing(
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            anyOrNull(),
+            eq(fixture.localMetricsAggregator)
         )
     }
 }